Moved pages and added content
