+++ src/routes/edit/+page.svelte
@@ -1,8 +1,156 @@
-<!-- Edit -->
 <script lang="ts">
+	// Core Components
+	import Button from '$lib/components/coreComponents/Button.svelte';
+	// Import Components
+	import Gamecard from '$lib/components/Gamecard.svelte';
+	import GamecardBack from '$lib/components/GamecardBack.svelte';
 	import Navbar from '$lib/components/Navbar.svelte';
+	import DiceIcon from '$lib/components/coreComponents/DiceIcon.svelte';
+
+	// Import Item Store
+	import { activeItem } from '$lib/stores/Items';
+	import ItemEditor from '$lib/components/ItemEditor.svelte';
+
+	import { items } from '$lib/stores/Items';
+	import { slide } from 'svelte/transition';
+
+	// Export to PDF section
+	var cardSection: HTMLElement;
+	async function printCards() {
+		window.print();
+	}
+
+	let editMode = true;
+
+	function newEmptyItem() {}
+	function toggleEditMode() {
+		editMode = !editMode;
+	}
 </script>
 
-<Navbar />
+<main id="main">
+	<!-- Header -->
+	<section id="header" class="hideOnPrint">
+		<Navbar />
+	</section>
+	<!-- Editor Pane -->
+	{#if editMode}
+		<section id="editor" class="hideOnPrint" transition:slide={{ duration: 200 }}>
+			<h2>Card Editor</h2>
+			<ItemEditor />
+			<div id="editorFooter">
+				<Button click={printCards} icon="mdi:printer">Print Card</Button>
+			</div>
+		</section>
+	{/if}
 
-<h1 class="text-3xl font-bold underline">Edit</h1>
+	<!-- Card Pane -->
+	<section id="cardView">
+		<div class="buttonRow hideOnPrint">
+			<Button click={toggleEditMode} icon={editMode ? 'mdi:eye' : 'mdi:pencil'}
+				>{editMode ? 'Viewing Mode' : 'Edit Card'}</Button
+			>
+			<Button click={newEmptyItem} icon="mdi:plus">New Card</Button>
+
+			<a class="mobileOnly" href="#editor">Go to editor</a>
+		</div>
+		<div id="cardArea" bind:this={cardSection}>
+			<Gamecard item={$activeItem} />
+			<GamecardBack item={$activeItem} />
+		</div>
+	</section>
+</main>
+
+<!-- Styles -->
+<style>
+	:global(body) {
+		margin: 0;
+	}
+	#main {
+		display: grid;
+		/* Section with scroll */
+		height: 100dvh;
+		overflow-y: hidden;
+		grid-template-areas: 'header header' 'cardView editor';
+		grid-template-columns: fit-content max(2fr, fit-content);
+		grid-template-rows: min-content 1fr;
+	}
+	@media print {
+		.hideOnPrint {
+			display: none !important;
+			visibility: hidden !important;
+			position: absolute;
+			top: 0;
+			left: 0;
+		}
+		.showOnPrint {
+			display: block;
+		}
+	}
+
+	/* Changed layout for mobile */
+	.mobileOnly {
+		display: none;
+	}
+	@media screen and (max-width: 750px) {
+		#main {
+			grid-template-areas: 'header' 'cardView' 'editor';
+			grid-template-rows: min-content min-content min-content;
+			grid-template-columns: 1fr;
+			height: fit-content;
+		}
+
+		.mobileOnly {
+			display: initial;
+		}
+	}
+
+	section {
+		padding: var(--padding);
+		/* Size and scroll */
+		overflow-y: auto;
+	}
+
+	section#header {
+		grid-area: header;
+	}
+
+	section#editor {
+		grid-area: editor;
+		background-color: aliceblue;
+		/* Layout */
+		display: flex;
+		flex-direction: column;
+		/* Size */
+		min-width: 20em;
+		box-sizing: border-box;
+		/* Scrollbar */
+		overflow-y: auto;
+		scroll-behavior: auto;
+	}
+
+	section#editor > #editorFooter {
+		margin-top: auto;
+		padding-top: var(--padding);
+		border-top: solid 1px black;
+	}
+
+	section#cardView {
+		/* Nothing to see here */
+	}
+
+	div#cardArea {
+		display: flex;
+		flex-wrap: wrap;
+		justify-content: center;
+		align-items: center;
+		/* Gap */
+		gap: 5mm;
+		padding: 5mm;
+	}
+
+	.buttonRow {
+		display: flex;
+		gap: var(--padding);
+	}
+</style>
